Refactored the entire project into a microservice architecture with new dependencies
