+++ Customer-Service/src/main/java/csci318/group10/customerservice/Controller/CustomerController.java
@@ -1,7 +1,6 @@
 package csci318.group10.customerservice.Controller;
 
 
-import csci318.group10.customerservice.Controller.DTO.CustomerDTO;
 import csci318.group10.customerservice.Model.Customer;
 import csci318.group10.customerservice.Service.CustomerService;
 import org.springframework.web.bind.annotation.*;
@@ -9,6 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
+@RequestMapping("/customer")
 public class CustomerController {
     private final CustomerService customerService;
 
@@ -16,12 +16,12 @@ public class CustomerController {
         this.customerService = customerService;
     }
 
-    @GetMapping("/customer")
+    @GetMapping("/all")
     List<Customer> allCustomers() {
         return customerService.getAllCustomers();
     }
 
-    @GetMapping("/customer/{id}")
+    @GetMapping("/{id}")
     Customer getCustomer(@PathVariable int id) {
         return customerService.getCustomerById(id);
     }
