--- Customer-Service/src/main/java/csci318/group10/customerservice/Controller/DTO/CustomerDTO.java
@@ -1,11 +1,0 @@
-package csci318.group10.customerservice.Controller.DTO;
-
-public class CustomerDTO {
-
-    private int id;
-    private String firstName;
-    private String lastName;
-
-    public CustomerDTO() {}
-
-}
