+++ Customer-Service/src/main/resources/application.properties
@@ -1,5 +1,4 @@
-spring.application.name=Customer-Service
-server.address=localhost
+spring.application.name=customer-service
 server.port=8081
 spring.h2.console.enabled=true
 spring.datasource.url=jdbc:h2:mem:custdb
