+++ Product-Service/src/main/resources/application.properties
@@ -1,5 +1,4 @@
-spring.application.name=Product-Service
-server.address=localhost
+spring.application.name=product-service
 server.port=8082
 spring.h2.console.enabled=true
-spring.datasource.url=jdbc:h2:mem:prodsdb
+spring.datasource.url=jdbc:h2:mem:productdb
