+++ Product-Service/src/test/java/csci318/group10/product_service/ProductServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package csci318.group10.productservice;
+package csci318.group10.product_service;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class ProductServiceApplicationTests {
 
-    @Test
-    void contextLoads() {
-    }
+	@Test
+	void contextLoads() {
+	}
 
 }
