+++ api-gateway/src/main/java/csci318/group10/api_gateway/ApiGatewayApplication.java
@@ -1,0 +1,13 @@
+package csci318.group10.api_gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApiGatewayApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApiGatewayApplication.class, args);
+	}
+
+}
