+++ api-gateway/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.application.name=api-gateway
+server.port=8080
+
+spring.cloud.gateway.discovery.locator.enabled=true
+spring.cloud.gateway.discovery.locator.lower-case-service-id=true
