+++ api-gateway/src/test/java/csci318/group10/api_gateway/ApiGatewayApplicationTests.java
@@ -1,0 +1,13 @@
+package csci318.group10.api_gateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiGatewayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
