+++ discovery-server/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.application.name=discovery-server
+server.port=8761
+
+eureka.client.register-with-eureka=false
+eureka.client.fetch-registry=false
