+++ discovery-server/src/test/java/csci318/group10/discovery_server/DiscoveryServerApplicationTests.java
@@ -1,0 +1,13 @@
+package csci318.group10.discovery_server;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DiscoveryServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
